+++ package/components/utils/CoreFade.js
@@ -18,8 +18,8 @@ import CoreMenu from "../navigation/CoreMenu";
 
 export default function CoreProfilePopOver(props) {
   const navigate = nativeUseNavigate();
-  const auth = useSelector((state) => state.auth);
-  const profile = useSelector((state) => state.profile);
+  const auth = useSelector((state) => state?.auth);
+  const profile = useSelector((state) => state?.profile);
   // eslint-disable-next-line no-unsafe-optional-chaining
   const { firstName = "", middleName = "", lastName = "" } = profile?.basic;
   const { onClose } = props;